Big changes to database, adding elections, and more robust code for processing votes and login. additional changes to editing and adding candidates and positons

+++ votesystem/admin/positions_delete.php
@@ -1,14 +1,22 @@
 <?php
+ini_set('display_errors', 1);
+ini_set('display_startup_errors', 1);
+error_reporting(E_ALL);
+
 	include 'includes/session.php';
 
 	if(isset($_POST['delete'])){
-		$id = $_POST['id'];
-		$sql = "DELETE FROM positions WHERE id = $1";
-		if(pg_query_params($conn, $sql, array($id))){
-			$_SESSION['success'] = 'Position deleted successfully';
-		}
-		else{
-			$_SESSION['error'] = pg_last_error($conn);
+		if (isset($_POST['id']) && is_numeric($_POST['id'])) {
+			$id = $_POST['id'];
+			$sql = "DELETE FROM positions WHERE id = $1";
+			if(pg_query_params($conn, $sql, array($id))){
+				$_SESSION['success'] = 'Position deleted successfully';
+			}
+			else{
+				$_SESSION['error'] = pg_last_error($conn);
+			}
+		} else {
+			$_SESSION['error'] = 'Invalid item to delete';
 		}
 	}
 	else{
